Code cleaning
Removed redundancy
