+++ app/src/main/java/com/iaspp/prepareyourself/movie/MovieResponseDTO.java
@@ -3,13 +3,17 @@ package com.iaspp.prepareyourself.movie;
 import com.google.gson.annotations.SerializedName;
 import com.iaspp.prepareyourself.interfaces.IDTO;
 
-public abstract class AbstractResponseDTO implements IDTO {
+import java.util.List;
+
+public class MovieResponseDTO implements IDTO {
 
     private int page;
     @SerializedName("total_pages")
     private int totalPages;
     @SerializedName("total_results")
     private int totalResults;
+    @SerializedName("results")
+    private List<MovieDTO> resultList;
 
     public int getPage() {
         return page;
@@ -35,17 +39,16 @@ public abstract class AbstractResponseDTO implements IDTO {
         this.totalResults = totalResults;
     }
 
+    public List<MovieDTO> getResultList() {
+        return resultList;
+    }
+
+    public void setResultList(List<MovieDTO> resultList) {
+        this.resultList = resultList;
+    }
+
     @Override
     public String toString() {
-        StringBuffer sb = new StringBuffer();
-        sb.append("Page: ");
-        sb.append(page);
-        sb.append("/");
-        sb.append(totalPages);
-        sb.append(" ");
-        sb.append(totalResults);
-        sb.append(" movies");
-
-        return sb.toString();
+        return "Page: " + page + "/" + totalPages + " " + totalResults + " movies " + " results: " + resultList;
     }
 }
