+++ app/src/main/res/layout/movie_upcoming.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.v7.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:cv="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:gravity="center_horizontal"
@@ -66,27 +67,37 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:gravity="center_horizontal"
-            android:orientation="horizontal">
+            android:orientation="horizontal"
+            android:padding="2dp">
 
             <Button
+                android:id="@+id/bt_link"
                 style="?attr/borderlessButtonStyle"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_weight="3" />
+                android:layout_weight="3"
+                android:adjustViewBounds="true"
+                android:scaleType="fitCenter"
+                android:text="@string/bt_link" />
 
             <Button
+                android:id="@+id/bt_share"
                 style="?attr/borderlessButtonStyle"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_weight="3" />
+                android:layout_weight="3"
+                android:scaleType="fitCenter"
+                android:text="@string/bt_share" />
 
             <Button
+                android:id="@+id/bt_view"
                 style="?attr/borderlessButtonStyle"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_weight="3" />
+                android:layout_weight="3"
+                android:scaleType="fitCenter"
+                android:text="@string/bt_view" />
         </LinearLayout>
     </LinearLayout>
-
 
 </android.support.v7.widget.CardView>
